requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
openai==1.30.1
Pillow==10.3.0
python-dotenv==1.0.1
gunicorn==22.0.0
werkzeug==3.0.3
